labs/knowledge_graphs/maps_recommender/requirements.txt
absl-py==2.2.2
alabaster==0.7.16
ampligraph==2.1.0
astunparse==1.6.3
babel==2.17.0
beautifultable==1.1.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contextlib2==21.6.0
contourpy==1.3.2
cycler==0.12.1
docopt==0.6.2
docutils==0.17.1
exceptiongroup==1.2.2
flake8==7.2.0
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
isodate==0.7.2
jinja2==3.1.6
joblib==1.4.2
keras==3.9.2
kiwisolver==1.4.8
latexcodec==3.0.0
libclang==18.1.1
markdown==3.8
markdown-it-py==2.2.0
markupsafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mdit-py-plugins==0.3.5
mdurl==0.1.2
ml-dtypes==0.5.1
mypy-extensions==1.0.0
myst-parser==0.18.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opt-einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
pip==25.0.1
platformdirs==4.3.7
pluggy==1.5.0
protobuf==5.29.4
pybtex==0.24.0
pybtex-docutils==1.0.3
pycodestyle==2.13.0
pyflakes==3.3.2
pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
rdflib==7.1.4
requests==2.32.3
rich==14.0.0
schema==0.7.5
scikit-learn==1.6.1
scipy==1.10.0
setuptools==78.1.0
six==1.17.0
snowballstemmer==2.2.0
sphinx==5.0.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.4.2
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uv==0.6.14
wcwidth==0.2.13
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
faiss-cpu==1.10.0
geopy==2.4.1
